Reverts earlier commit
We need to have the code be independant of transpilation, not locked into it
like this commit proposes. Test runners need to do on the fly transpilation
and work off source.

The direct consequence of the commit I am reverting is that during flow-type
checks, tests are evaluated against raw untyped javascript - we'll never get
our coverage to rise this way.

This reverts commit f15ad2c89a2a99f694bb7e1170b5788b599710a0.